Add first test for induced_distribution module
	Function tested: genrate_sample_measure
	Used to generate the correct function method
	for induced sampling

	Move previous induced distribution code
	to development/induced_distribution_legacy.py
