+++ tests/test_induced_distributions.py
@@ -1,0 +1,11 @@
+from unittest import TestCase
+from equadratures.induced_distributions import InducedSampling
+
+
+class TestInducedDistribution(TestCase):
+
+    def test_generate_sample_measure(self):
+        # test if the method returns a function object for induced sampling
+        measure = InducedSampling(1, 1, 1, "Chebyshev", 0, 0)
+        func = measure.generate_sample_measure("Chebyshev", 0, 0)
+        assert callable(func)
